Greatly simplify the cognito identity provider
This change adds a storage module for CognitoUserPool which means
that we can use the normal functions from the UserPool object to
manage the CognitoUser and CognitoUserSession objects instead of
building them manually.

This should remove a lot of the issues we have had in the past with
manually attempting to maintain the cognito session.